# spec/openapi.yaml
openapi: 3.1.0
info:
  title: MAIL Server API
  version: "2.0-pre1"
  description: REST API for the Multi-Agent Interface Layer (MAIL) server, including interswarm messaging.
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    MAILMessage:
      $ref: './MAIL-core.schema.json#/$defs/MAILMessage'
...
    MAILServerSentEvent:
      $ref: './MAIL-core.schema.json#/$defs/MAILServerSentEvent'
paths:
...
  /message:
    post:
      summary: Send a message to the swarm 
      description: Requires role=user or role=admin.
      operationId: message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [message]
              properties:
                message: {
                  $ref: '#/components/schemas/MAILMessage'
                }
                entrypoint:
                  type: string
                  description: Optional agent entrypoint override (defaults to configured entrypoint)
                show_events:
                  type: boolean
                  description: When true, include collected events alongside the response
                  default: false
                stream:
                  type: boolean
                  description: When true, server responds with text/event-stream (SSE)
                  default: false
                resume_from:
                  type: string
                  description: When provided, resume the task from the given state
                  enum: ["user_response", "breakpoint_tool_call"]
                  default: null
                kwargs:
                  type: object
                  description: Additional keyword arguments to pass to the entrypoint
                  default: {}
      responses:
        '200':
          description: Final response produced by supervisor after task completion
          content:
            application/json:
              schema:
                type: object
                required: [response]
                additionalProperties: false
                properties:
                  response: { 
                    $ref: '#/components/schemas/MAILMessage' 
                  }
                  events:
                    type: array
                    nullable: true
                    items:
                      $ref: '#/components/schemas/MAILServerSentEvent'
            text/event-stream:
              schema:
                type: string
                description: SSE stream emitted when stream=true
        '401':
          description: Unauthorized
...